requirements.txt
pandas~=2.3.1
python-dotenv~=1.1.1
selenium~=4.35.0
wget~=3.2